Add files via upload
I found this assignment even easier then the first.  Maybe a little bit because it was an easier thing to construct, but mainly because i have just become much more comfortable with OnShape.  I am having a great time doing these projects and I look forward to the next part of the caster.